sample_benchmark/requirements.txt
absl-py==1.1.0
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backoff==2.2.1
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.11.2
bleach==6.0.0
blessed==1.20.0
CacheControl==0.12.10
cachetools==5.2.0
cachy==0.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cleo==0.8.1
click==8.1.3
clikit==0.6.2
cloudpickle==2.2.1
colorful==0.5.5
comm==0.1.2
crashtest==0.3.1
dask==2023.2.1
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
distributed==2023.2.1
dm-tree==0.1.8
entrypoints==0.4
executing==1.2.0
fastapi==0.92.0
fastjsonschema==2.16.2
filelock==3.9.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
google-api-core==2.11.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.58.0
gpustat==1.0.0
grpcio==1.47.0
Gymnasium==0.26.3
gymnasium-notices==0.0.1
h11==0.14.0
HeapDict==1.0.1
idna==3.4
imageio==2.25.1
importlib-metadata==4.12.0
importlib-resources==5.12.0
ipykernel==6.19.2
ipyparallel==8.4.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
joblibspark==0.5.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==7.4.8
jupyter-console==6.6.1
jupyter_core==5.1.0
jupyter-events==0.6.3
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keyring==21.8.0
kubernetes==26.1.0
locket==1.0.0
lockfile==0.12.2
lz4==4.3.2
Markdown==3.3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
msgpack==1.0.4
multidict==6.0.4
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
nvidia-ml-py==11.495.46
oauthlib==3.2.2
opencensus==0.11.1
opencensus-context==0.1.3
opentelemetry-api==1.16.0
opentelemetry-exporter-otlp==1.16.0
opentelemetry-exporter-otlp-proto-grpc==1.16.0
opentelemetry-exporter-otlp-proto-http==1.16.0
opentelemetry-proto==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
platformdirs==2.6.0
poetry==1.1.12
poetry-core==1.0.7
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
py4j==0.10.9.5
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.5
Pygments==2.14.0
pylev==1.4.0
pyrsistent==0.19.3
pyspark==3.3.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
ray==2.3.0
ray-cpp==2.3.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.1
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.1
Send2Trash==1.8.0
setuptools==65.6.3
shellingham==1.4.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.25.0
tabulate==0.9.0
tblib==1.7.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.2.3
tinycss2==1.2.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
tune-sklearn==0.4.5
typer==0.7.0
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.19.0
watermark==2.3.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
wheel==0.38.4
widgetsnbextension==4.0.5
wrapt==1.14.1
yarl==1.8.2
zict==2.2.0
zipp==3.8.0
